+++ lib/src/build.dart
@@ -1,7 +1,6 @@
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/material.dart' as prefix0;
 
 import 'model.dart';
 
@@ -255,14 +254,16 @@ class _TxtBuildEditableState extends State<TxtBuildEditable> {
       _placeholderStyle = TextStyle(
         fontWeight: widget.textModel?.textStyle?.fontWeight,
         fontSize: widget.textModel?.textStyle?.fontSize,
-        color: widget.textModel?.textStyle?.color?.withOpacity(0.7) ?? Colors.grey,
+        color:
+            widget.textModel?.textStyle?.color?.withOpacity(0.7) ?? Colors.grey,
         fontStyle: widget.textModel?.textStyle?.fontStyle ?? FontStyle.normal,
         fontFamily: widget.textModel?.textStyle?.fontFamily,
         fontFamilyFallback: widget.textModel?.textStyle?.fontFamilyFallback,
         letterSpacing: widget.textModel?.textStyle?.letterSpacing,
         wordSpacing: widget.textModel?.textStyle?.wordSpacing,
       );
-      _placehodlerController = TextEditingController(text: widget.textModel?.placeholder);
+      _placehodlerController =
+          TextEditingController(text: widget.textModel?.placeholder);
     }
 
     return EditableText(